// Furniture-Store-Management-FE-main/client/src/entities/RepairOrder.ts
type RepairOrder = {
  id: number;
  productName: string;
  description: string;
  details: string;
  cost: string;
  status: string;
  estimateFinishDate: string;
  finishDate: string;
  createdAt: string;
  updatedAt: string;
  customerId: number;
  staffId: number;
};
export default RepairOrder;